fix: Revert moving createIcons to separate file
